+++ app/controllers/hr_lite/admin/salary_structures_controller.rb
@@ -1,6 +1,6 @@
 module HrLite
   module Admin
-    class SalaryStructuresController < LeadershipController
+    class SalaryStructuresController < SuperadminController
       before_action :set_profile
 
       def new