+++ app/controllers/hr_lite/admin/settings_controller.rb
@@ -7,7 +7,7 @@ module HrLite
 
       def update
         @setting = Setting.instance
-        if @setting.update(params.require(:setting).permit(:weekend_policy))
+        if @setting.update(params.require(:setting).permit(:weekend_policy, :employee_code_prefix))
           redirect_to edit_admin_setting_path, notice: "Settings saved."
         else
           render :edit, status: :unprocessable_entity